+++ src/pipeline/diff_lister.rs
@@ -2,7 +2,7 @@ use std::collections::HashSet;
 
 use anyhow::Result;
 use aws_sdk_s3::types::Object;
-use tracing::{debug, trace};
+use tracing::{debug, error, trace};
 
 use crate::types::{ObjectKey, ObjectKeyMap, S3syncObject};
 
@@ -22,6 +22,19 @@ impl DiffLister {
         source_key_map: &ObjectKeyMap,
         target_key_map: &ObjectKeyMap,
     ) -> Result<()> {
+        // This is special for test emulation.
+        #[allow(clippy::collapsible_if)]
+        if cfg!(feature = "e2e_test_dangerous_simulations") {
+            panic_simulation(&self.base.config, "DiffLister::receive_and_filter");
+
+            if is_error_simulation_point(&self.base.config, "DiffLister::receive_and_filter") {
+                error!("error simulation point has been triggered.");
+                return Err(anyhow::anyhow!(
+                    "error simulation point has been triggered."
+                ));
+            }
+        }
+
         trace!("diff generator has started.");
 
         let diff_set = generate_diff(source_key_map, target_key_map);
@@ -98,6 +111,37 @@ fn generate_diff(source_key_map: &ObjectKeyMap, target_key_map: &ObjectKeyMap) -
     }
 
     diff_set
+}
+
+#[cfg_attr(coverage_nightly, coverage(off))]
+fn panic_simulation(config: &crate::Config, panic_simulation_point: &str) {
+    const PANIC_DANGEROUS_SIMULATION_ENV: &str = "S3SYNC_PANIC_DANGEROUS_SIMULATION";
+    const PANIC_DANGEROUS_SIMULATION_ENV_ALLOW: &str = "ALLOW";
+
+    if std::env::var(PANIC_DANGEROUS_SIMULATION_ENV)
+        .is_ok_and(|v| v == PANIC_DANGEROUS_SIMULATION_ENV_ALLOW)
+        && config
+            .panic_simulation_point
+            .as_ref()
+            .is_some_and(|point| point == panic_simulation_point)
+    {
+        panic!(
+            "panic simulation has been triggered. This message should not be shown in the production.",
+        );
+    }
+}
+
+#[cfg_attr(coverage_nightly, coverage(off))]
+fn is_error_simulation_point(config: &crate::Config, error_simulation_point: &str) -> bool {
+    const ERROR_DANGEROUS_SIMULATION_ENV: &str = "S3SYNC_ERROR_DANGEROUS_SIMULATION";
+    const ERROR_DANGEROUS_SIMULATION_ENV_ALLOW: &str = "ALLOW";
+
+    std::env::var(ERROR_DANGEROUS_SIMULATION_ENV)
+        .is_ok_and(|v| v == ERROR_DANGEROUS_SIMULATION_ENV_ALLOW)
+        && config
+            .error_simulation_point
+            .as_ref()
+            .is_some_and(|point| point == error_simulation_point)
 }
 
 #[cfg(test)]