+++ src/storage/checksum/crc32.rs
@@ -148,6 +148,7 @@ mod tests {
         assert_eq!(checksum.finalize_all(), FULL_OBJECT_CHECKSUM.to_string());
     }
 
+    #[cfg_attr(coverage_nightly, coverage(off))]
     async fn create_large_file() {
         if PathBuf::from(LARGE_FILE_PATH).try_exists().unwrap() {
             return;