+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.44.0] - 2025-09-12
+
+### Changed
+- Added dry_run info to an event callback
+- Improved error handling in async functions  
+  Even if an async function fails with Rust panic, s3sync can handle it as an error.
+- Updated dependencies
+- Updated docs
+
 ## [1.43.1] - 2025-09-11
 
 ### Changed